# shared/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/schedule/": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "summary": "Get current schedule",
        "description": "Retrieves the full schedule using the GetSchedule interactor.",
        "operationId": "get_schedule_schedule__get",
        "responses": {
          "200": {
            "description": "Schedule retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetScheduleOutput"
                }
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/schedule/subscriptions/": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "summary": "List the current user's subscriptions",
        "description": "Returns every schedule event the current user is subscribed to. Served separately from the schedule so the schedule itself stays universal and cacheable offline.",
        "operationId": "get_subscriptions_schedule_subscriptions__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSubscriptionsOutput"
                }
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "Schedule"
        ],
        "summary": "Create a new event subscription",
        "description": "Subscribes the current user to a specific schedule event. Prevents duplicate subscriptions.",
        "operationId": "new_subscription_schedule_subscriptions__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Subscription created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSubscriptionOutput"
                }
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Event ID does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "409": {
            "description": "Subscription for this event already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/schedule/subscriptions/{subscription_id}": {
      "delete": {
        "tags": [
          "Schedule"
        ],
        "summary": "Remove a subscription",
        "description": "Deletes an existing subscription by its unique ID.",
        "operationId": "delete_subscription_schedule_subscriptions__subscription_id__delete",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Subscription ID to remove.",
              "title": "Subscription Id"
            },
            "description": "Subscription ID to remove."
          }
        ],
        "responses": {
          "204": {
            "description": "Subscription deleted successfully."
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Subscription ID not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
...
    "/voting/status": {
      "get": {
        "tags": [
          "Voting"
        ],
        "summary": "Get current voting state",
        "description": "Retrieves the current phase of the voting process (e.g., active, closed) and reasoning.",
        "operationId": "get_voting_status_voting_status_get",
        "responses": {
          "200": {
            "description": "Voting status retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVotingStateOutput"
                }
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/voting/nominations": {
      "get": {
        "tags": [
          "Voting"
        ],
        "summary": "List all nominations for the current vote",
        "description": "Retrieves a list of all candidates or items eligible for voting in the current session.",
        "operationId": "list_voting_nominations_voting_nominations_get",
        "responses": {
          "200": {
            "description": "Nominations retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListVotingNominationsOutput"
                }
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/voting/nominations/{nomination_code}": {
      "get": {
        "tags": [
          "Voting"
        ],
        "summary": "Get voting nomination details",
        "description": "Retrieves detailed information about a specific nomination.",
        "operationId": "get_voting_nomination_voting_nominations__nomination_code__get",
        "parameters": [
          {
            "name": "nomination_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Voting nomination code.",
              "title": "Nomination Code"
            },
            "description": "Voting nomination code."
          }
        ],
        "responses": {
          "200": {
            "description": "Nomination details retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetVotingNominationOutput"
                }
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Nomination not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
    "/voting/votes": {
      "post": {
        "tags": [
          "Voting"
        ],
        "summary": "Cast a vote",
        "description": "Submits a vote for the given participant. The nomination is derived from the participant.",
        "operationId": "add_vote_voting_votes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddVoteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Vote successfully cast.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddVoteOutput"
                }
              }
            }
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "Participant not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "409": {
            "description": "Already voted in this nomination.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/voting/votes/{vote_id}": {
      "delete": {
        "tags": [
          "Voting"
        ],
        "summary": "Cancel a vote",
        "description": "Removes a previously cast vote by its ID.",
        "operationId": "cancel_vote_voting_votes__vote_id__delete",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "vote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "ID of the vote to cancel.",
              "title": "Vote Id"
            },
            "description": "ID of the vote to cancel."
          }
        ],
        "responses": {
          "204": {
            "description": "Vote successfully cancelled."
          },
          "422": {
            "description": "Request validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "403": {
            "description": "Access denied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          },
          "404": {
            "description": "No vote found to cancel.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMessage"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddVoteInput": {
        "properties": {
          "participant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Participant Id"
          }
        },
        "type": "object",
        "required": [
          "participant_id"
        ],
        "title": "AddVoteInput"
      },
      "AddVoteOutput": {
        "properties": {
          "vote_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vote Id"
          }
        },
        "type": "object",
        "required": [
          "vote_id"
        ],
        "title": "AddVoteOutput"
      },
...
      "CreateSubscriptionInput": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id"
          },
          "counter": {
            "type": "integer",
            "title": "Counter"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "counter"
        ],
        "title": "CreateSubscriptionInput"
      },
      "CreateSubscriptionOutput": {
        "properties": {
          "subscription_id": {
            "type": "string",
            "format": "uuid",
            "title": "Subscription Id"
          }
        },
        "type": "object",
        "required": [
          "subscription_id"
        ],
        "title": "CreateSubscriptionOutput"
      },
...
      "ErrorMessage": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ErrorMessage"
      },
      "GetScheduleOutput": {
        "properties": {
          "schedule": {
            "items": {
              "$ref": "#/components/schemas/ScheduleEventFullDTO"
            },
            "type": "array",
            "title": "Schedule"
          }
        },
        "type": "object",
        "required": [
          "schedule"
        ],
        "title": "GetScheduleOutput"
      },
      "GetSubscriptionsOutput": {
        "properties": {
          "subscriptions": {
            "items": {
              "$ref": "#/components/schemas/SubscriptionFullDTO"
            },
            "type": "array",
            "title": "Subscriptions"
          }
        },
        "type": "object",
        "required": [
          "subscriptions"
        ],
        "title": "GetSubscriptionsOutput"
      },
      "GetVotingNominationOutput": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "works_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Works Url"
          },
          "participants_count": {
            "type": "integer",
            "title": "Participants Count"
          },
          "user_vote": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NominationVoteDTO"
              },
              {
                "type": "null"
              }
            ]
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/ParticipantFullDTO"
            },
            "type": "array",
            "title": "Participants"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "title",
          "participants_count",
          "user_vote",
          "participants"
        ],
        "title": "GetVotingNominationOutput"
      },
      "GetVotingStateOutput": {
        "properties": {
          "can_vote": {
            "type": "boolean",
            "title": "Can Vote"
          },
          "status": {
            "$ref": "#/components/schemas/VotingStatus"
          }
        },
        "type": "object",
        "required": [
          "can_vote",
          "status"
        ],
        "title": "GetVotingStateOutput"
      },
...
      "ListVotingNominationsOutput": {
        "properties": {
          "nominations": {
            "items": {
              "$ref": "#/components/schemas/NominationVotingDTO"
            },
            "type": "array",
            "title": "Nominations"
          }
        },
        "type": "object",
        "required": [
          "nominations"
        ],
        "title": "ListVotingNominationsOutput"
      },
...
      "NominationVoteDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "participant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Participant Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "participant_id"
        ],
        "title": "NominationVoteDTO"
      },
      "NominationVotingDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "works_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Works Url"
          },
          "participants_count": {
            "type": "integer",
            "title": "Participants Count"
          },
          "user_vote": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NominationVoteDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "title",
          "participants_count",
          "user_vote"
        ],
        "title": "NominationVotingDTO"
      },
...
      "ParticipantFullDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "voting_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voting Number"
          },
          "votes_count": {
            "type": "integer",
            "title": "Votes Count"
          },
          "user_vote": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParticipantVoteDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "voting_number",
          "votes_count",
          "user_vote"
        ],
        "title": "ParticipantFullDTO"
      },
      "ParticipantVoteDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "ParticipantVoteDTO"
      },
...
      "ScheduleEventFullDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "duration": {
            "type": "integer",
            "title": "Duration"
          },
          "order": {
            "type": "number",
            "title": "Order"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current"
          },
          "is_skipped": {
            "type": "boolean",
            "title": "Is Skipped"
          },
          "nomination_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nomination Title"
          },
          "block_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Title"
          },
          "queue": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue"
          },
          "time_until": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Until"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "title",
          "duration",
          "order",
          "is_current",
          "is_skipped",
          "nomination_title",
          "block_title",
          "queue",
          "time_until"
        ],
        "title": "ScheduleEventFullDTO"
      },
...
      "SubscriptionEventDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "order": {
            "type": "number",
            "title": "Order"
          },
          "queue": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue"
          },
          "time_until": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Until"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "title",
          "order",
          "queue",
          "time_until"
        ],
        "title": "SubscriptionEventDTO"
      },
      "SubscriptionFullDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "counter": {
            "type": "integer",
            "title": "Counter"
          },
          "event": {
            "$ref": "#/components/schemas/SubscriptionEventDTO"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "counter",
          "event"
        ],
        "title": "SubscriptionFullDTO"
      },
...
      "ValidationErrorDetail": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "type"
        ],
        "title": "ValidationErrorDetail"
      },
      "ValidationErrorResponse": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "default": "VALIDATION_ERROR"
          },
          "details": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ValidationErrorDetail"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "title": "ValidationErrorResponse"
      },
      "VotingStatus": {
        "type": "string",
        "enum": [
          "open",
          "not_authenticated",
          "no_ticket",
          "disabled"
        ],
        "title": "VotingStatus"
      }
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "Session cookie set after login.",
        "in": "cookie",
        "name": "session_id"
      }
    }
  }
}